Relocate config to core
