--- src/app/config/__init__.py
@@ -1,7 +1,0 @@
-__all__ = [
-    'Config',
-    'get_config',
-]
-
-from config.base import Config
-from config.dependencies import get_config
