+++ src/app/core/config/__init__.py
@@ -1,0 +1,7 @@
+__all__ = [
+    'Config',
+    'get_config',
+]
+
+from core.config.base import Config
+from core.config.dependencies import get_config
