+++ src/app/core/config/development.py
@@ -1,6 +1,6 @@
 __all__ = ['DevelopmentConfig']
 
-from config.base import BaseConfig
+from core.config.base import BaseConfig
 
 
 class DevelopmentConfig(BaseConfig):
