+++ src/app/core/config/testing.py
@@ -1,6 +1,6 @@
 __all__ = ['TestingConfig']
 
-from config.base import BaseConfig
+from core.config.base import BaseConfig
 
 
 class TestingConfig(BaseConfig):
