+++ src/app/core/exceptions.py
