+++ src/tests/unit/config/test_dependencies.py
@@ -1,8 +1,8 @@
 import pytest
 from pytest_mock import MockerFixture
 
-from config import Config, get_config
-from config.exceptions import ImproperlyConfigured
+from core.config import Config, get_config
+from core.exceptions import ImproperlyConfigured
 from tests.unit.config.mocks import ConfigMock
 
 
@@ -10,7 +10,7 @@ class TestDependencies:
     def test_get_config_with_environ_and_config_in_registry(self, mocker: MockerFixture) -> None:
         # Arrange
         expected: Config = ConfigMock()
-        mocker.patch('config.dependencies._get_config_registry', return_value={'mock': ConfigMock})
+        mocker.patch('core.config.dependencies._get_config_registry', return_value={'mock': ConfigMock})
         mocker.patch('os.environ', {'FASTAPI_CONFIGURATION': 'MOCK'})
 
         # Act
@@ -21,7 +21,7 @@ class TestDependencies:
 
     def test_get_config_without_environ_and_config_in_registry(self, mocker: MockerFixture) -> None:
         # Arrange
-        mocker.patch('config.dependencies._get_config_registry', return_value={'mock': ConfigMock})
+        mocker.patch('core.config.dependencies._get_config_registry', return_value={'mock': ConfigMock})
         mocker.patch('os.environ', {})
 
         # Act + Assert
@@ -30,7 +30,7 @@ class TestDependencies:
 
     def test_get_config_with_environ_and_config_not_in_registry(self, mocker: MockerFixture) -> None:
         # Arrange
-        mocker.patch('config.dependencies._get_config_registry', return_value={})
+        mocker.patch('core.config.dependencies._get_config_registry', return_value={})
         mocker.patch('os.environ', {'FASTAPI_CONFIGURATION': 'mock'})
 
         # Act + Assert
@@ -41,7 +41,7 @@ class TestDependencies:
     # We already tested behaviour with and without environ or with and without configuration in registry
     def test_get_config_without_environ_and_config_not_in_registry(self, mocker: MockerFixture) -> None:
         # Arrange
-        mocker.patch('config.dependencies._get_config_registry', return_value={})
+        mocker.patch('core.config.dependencies._get_config_registry', return_value={})
         mocker.patch('os.environ', {})
 
         # Act + Assert
